Add eventType to Appointment model and completed to Event model

+++ ApPoint/app/src/main/java/com/bachelor/appoint/model/Risk.kt
@@ -2,5 +2,5 @@ package com.bachelor.appoint.model
 
 class Risk(
     val type: String = "",
-    val value: Int = 1
+    val value: Double = 1.0
 )
